fix: Ask CEF to not relaunch the process if running as admin
CEF has a security feature that attempts to de-elevate when running as
admin. This improves user security in case of exploits. However, CEF
relaunches the "outer" process (e.g. our main process), not only the
browser subprocess. We're not equipped to handle unannounced hidden
process restarts for several reasons:

* CEF passes all of its configuration as command line arguments to the
outer process, which doesn't understand them
* We may attempt to hold locks or resources to files (such as log files)
multiple times, as we load CEF quite late in the init loop. CEF does not
forcefully terminate the "old" process when relaunching.

This relaunch design appears nontrivial to support from our perspective.
The easy fix is to disable the de-elevation feature. This isn't great
for user security unfortunately.